+++ rbac-distributed-service-impl/src/test/java/com/dwarfeng/rbacds/impl/service/PermissionGroupMaintainServiceImplTest.java
@@ -15,6 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -69,9 +70,14 @@ public class PermissionGroupMaintainServiceImplTest {
             assertEquals(permissionGroups.size(), lookup.getCount());
         } finally {
             for (PermissionGroup permissionGroup : permissionGroups) {
+                if (Objects.isNull(permissionGroup.getKey())) {
+                    continue;
+                }
                 permissionGroupMaintainService.deleteIfExists(permissionGroup.getKey());
             }
-            permissionGroupMaintainService.deleteIfExists(parentPermissionGroup.getKey());
+            if (Objects.nonNull(parentPermissionGroup.getKey())) {
+                permissionGroupMaintainService.deleteIfExists(parentPermissionGroup.getKey());
+            }
         }
     }
 }